requirements.txt
feedparser
requests
beautifulsoup4
python-dateutil
python-dotenv
supabase
PyYAML
lxml
html5lib